Resolve "HU10-services"

Closes #2

See merge request acortinas/asi-2122-rest-example!2

+++ src/main/java/es/udc/asi/restexample/model/service/PostService.java
@@ -1,5 +1,6 @@
 package es.udc.asi.restexample.model.service;
 
+import java.time.LocalDateTime;
 import java.util.List;
 import java.util.stream.Collectors;
 
@@ -10,15 +11,16 @@ import org.springframework.transaction.annotation.Transactional;
 
 import es.udc.asi.restexample.model.domain.Post;
 import es.udc.asi.restexample.model.exception.NotFoundException;
+import es.udc.asi.restexample.model.exception.OperationNotAllowed;
 import es.udc.asi.restexample.model.repository.PostDao;
 import es.udc.asi.restexample.model.repository.TagDao;
 import es.udc.asi.restexample.model.repository.UserDao;
 import es.udc.asi.restexample.model.service.dto.PostDTO;
+import es.udc.asi.restexample.model.service.dto.UserDTOPrivate;
 
 @Service
 @Transactional(readOnly = true, rollbackFor = Exception.class)
 public class PostService {
-
   @Autowired
   private PostDao postDAO;
 
@@ -27,6 +29,9 @@ public class PostService {
 
   @Autowired
   private TagDao tagDAO;
+
+  @Autowired
+  private UserService userService;
 
   public List<PostDTO> findAll() {
     return postDAO.findAll().stream().map(post -> new PostDTO(post)).collect(Collectors.toList());
@@ -42,7 +47,7 @@ public class PostService {
 
   // Con estas anotaciones evitamos que usuarios no autorizados accedan a ciertas
   // funcionalidades
-  @PreAuthorize("hasAuthority('ADMIN')")
+  @PreAuthorize("isAuthenticated()")
   @Transactional(readOnly = false)
   public PostDTO create(PostDTO post) {
     Post bdPost = new Post(post.getTitle(), post.getBody());
@@ -71,9 +76,24 @@ public class PostService {
     return new PostDTO(bdPost);
   }
 
-  @PreAuthorize("hasAuthority('ADMIN')")
+  @PreAuthorize("isAuthenticated()")
   @Transactional(readOnly = false)
-  public void deleteById(Long id) {
+  public void deleteById(Long id) throws NotFoundException, OperationNotAllowed {
+    Post post = postDAO.findById(id);
+    if (post == null) {
+      throw new NotFoundException(id.toString(), Post.class);
+    }
+
+    UserDTOPrivate currentUser = userService.getCurrentUserWithAuthority();
+    if (!currentUser.getId().equals(post.getAuthor().getId())) {
+      throw new OperationNotAllowed("Current user is not the post creator");
+    }
+    
+    LocalDateTime halfAnHourAgo = LocalDateTime.now().minusMinutes(30);
+    if (post.getTimestamp().isBefore(halfAnHourAgo)) {
+      throw new OperationNotAllowed("More than half an hour has passed since the post creation");
+    }
+
     postDAO.deleteById(id);
   }
 }
